more explicit command line option for normalization
a normalization was always applied linear or log  on whole matrix or by
row. but linear on whole matrix was the default and apply when option
was ommited.
to be more transparent add value 'lin' in accepted values for --norm.
'lin' is the default so now the normalization is more explicit.
but the behavior is the same.

fix the name of normalisation function ccording new option values
fix unit test according new function names
fix documentation according new api